feat(faculty)!: enhance faculty endpoints
Pagination, sorting and projection were added to faculty endpoints.

Furthermore, ModelFactory was refactored to permit even more
reusability. The rest of the models will be refactored to use
the new improved factory.

BREAKING CHANGE: Collection endpoints no longer err when zero items
are found. They now return a 200 status code and a 'totalCount'
field indicating the number of items found in the collection.